R2S1_P2/pyproject.toml
[project]
name = "projekt2"
version = "0.1.0"
description = "Projekt numer 2 (Automatyka) TMiPA 2024/2025"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "control>=0.10.2",
    "fpdf2>=2.8.5",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "sympy>=1.14.0",
]
